refactor: extract sidebar keyboard shortcuts and fix collapse behavior
Extract keyboard shortcut handler into sidebar/keyboard.ts module.
Fix closeDrawer to properly set collapsed state and persist to
localStorage so backdrop click and swipe close match toggle behavior.
